+++ src/app/app.component.ts
@@ -5,6 +5,7 @@
 
 // Services to inject
 import { DynamicHeadService } from './shared/services/dynamic-head.service'
+import { I18nService } from './shared/services/i18n.service'
 
 /**
  * Component name
@@ -17,7 +18,8 @@ export const appComponentName = 'appRoot'
 export class AppController {
 
   // Use the $inject property to ensure proper functionality after minification
-  public static $inject = ['$log', 'constants', 'dynamicHeadService']
+  public static $inject = ['$log', 'constants', 'dynamicHeadService',
+    'i18nService']
 
   /**
    * Create a controller
@@ -25,9 +27,11 @@ export class AppController {
    * @param {Object} constants - to log constants object in some environments.
    * @param {DynamicHeadService} dynamicHeadService - to add listeners for
    *  this component.
+   * @param {I18nService} i18nService - to load i18n text.
    */
   constructor(private $log: ng.ILogService, private constants,
-    private dynamicHeadService: DynamicHeadService) {
+    private dynamicHeadService: DynamicHeadService,
+    private i18nService: I18nService) {
 
   }
 
@@ -35,6 +39,10 @@ export class AppController {
    * Component initialization
    */
   $onInit() {
+    // Get i18n texts and load dynamic head elements.
+    this.i18nService.load()
+      .then(() => this.dynamicHeadService.update())
+
     // Log constants object
     this.$log.debug('Constants object', this.constants)
   }
@@ -43,8 +51,8 @@ export class AppController {
    * When component is linked
    */
   $postLink() {
-    // Start listener for route success to change title and meta tags
-    this.dynamicHeadService.start()
+    // Start listener for dynamicHeadService
+    this.dynamicHeadService.create()
   }
 }