+++ src/app/app.d.ts
@@ -1,5 +1,0 @@
-// Support AMD require for locale-en.json
-declare module 'src/assets/i18n/locale-en.json' {
-  const html: string
-  export default html
-}