+++ src/app/shared/services/services.module.ts
@@ -9,6 +9,7 @@
 
 // Shared services
 import { DynamicHeadService } from './dynamic-head.service'
+import { I18nService } from './i18n.service'
 import { StylesService } from './styles.service'
 
 /**
@@ -20,4 +21,5 @@ export const servicesModule = 'services'
 angular
   .module(servicesModule, [])
   .service(DynamicHeadService.serviceName, DynamicHeadService)
+  .service(I18nService.serviceName, I18nService)
   .service(StylesService.serviceName, StylesService)